Complete Project 9
Complete project 9, which augmented project 7 with GCD. At
first, we went over DispatchQueue and how you can use
various quality of servies on a global background thread to
prioritize work. Later on we implemented the performSelector
method that allowed us to not have to worry about strong
reference cycles or closures.