Finish more of project 11
Finished more of project 11. I added in spinning “good” and
“bad” slots to indicate which slot scores points, added physics
to the slots to determine when to destroy the balls, and also
added in a scoring mechanism. I also started to add an edit
mode, which will be completed in the next commit.

The main things I learned in this commit were how to use
SKAction to perform an “animation” on the screen as well as
how to add physics to a SKSpriteNode with physics bodies.