FInish Project 11
Finished Project 11 along with all but one of the challenges. I
learned a lot about SpriteKit’s various capabilities and got to
create a functional game with it (although there is no goal
right now). The one challenge I didn’t do of adding a goal to
the game could pretty easily be done by removing each “goal”
from the game after it is detected as hit and keeping track of
a ball count that has a property observer to watch for
the game over state. I would also have to create a list of goals
to determine when I have won the game. Regardless, just want
to move onto the next project :).