Finish Project 12A and 12B
Finished Project12A and Project 12B. These projects intro’d
the idea of UserDefaults, which is a mechanism for storing info
and retrieving information when the app first launches. The
project implements NSCoding in project 12A and Codable in
project 12B to show how to implement the storage and
retrieval for both methods (NSCoding = ObjC + Swift while
Codable is for Swift only).