Complete Project 13
Completed Project 13, which was an introduction to UISlider
(a pretty simple UI element that is a slider with a value) and
CoreImage (the very advanced utility for editing images).
It was pretty interesting to see just how complex CoreImage
could get, but it is really fast and makes the app look great.
We also touched on getting media from photos with
UIImagePicker and saving media to the users library with
UIImageWriteToSavedPhotosAlbum.