Complete Project 15
Completed Project 15, which was all about using animations
with Core Animation. We used the transform property of an
image view to do all sorts of things along with editing the
method parameters of UIView.animate() to achieve different
types of animations. It was a pretty short project, but pretty
interesting to see how expansive you can really be with
animation in iOS.