Complete Project 16
Completed project 16, which was a foray into Safari extensions
and JavaScript injection. It was kind of a weird project, but I
definately want to go back and do the challenges since they
invlove using User Defaults and I haven’t worked much with
them. I probably wouldn’t do the TableView challenge because
it is a little too involved and I want to move on.