Complete challenge one in Project 16
Completed the UserDefaults challenge in Project 16. I did a
pretty minimal copy job for the save function and the loading
code (I had forgotten how interacting with the JSON encoder
and decoder worked exactly). I had issues with trying to save
a pure string (which I assume is because it is trying to turn one
string into JSON), so I ended up saving the result as a string
array (which my best guess… works well with JSONEncoder).