Commit more of Project 17
Commit a little bit more of project 17. This commit added the
actual drawing of our slices along with playing a sound when
the slice is drawn to the screen. This project is super long, but
I am hoping to make a push tomorrow to get a large chunk
done and finish it off for good.