Complete Project 17
Completed Project 17, which was a lot more SpriteKit. The
components I felt like I could take to other projects were the
use of AVFoundation to play a sound (and stop whenever) as
well as potentially UIBezierPath, which was used to draw lines
based on a collection of CGPoints. I also utilized CaseIterable
for the first time, which allowed me to turn my enum into an
array of all of the enum options with the allCases method.