Complete Project 19
Completed proejct 19 with minimal implemtations of the two
challenges. The first challenge is a pretty mimimal by offering
three map options to choose from in the nav bar, while the
second challenge is a bit more involved. I am also not sure I
did it in the most efficient way by removing and readding the
annotation, but I couldn’t find an annotation redraw function
in my quick search. Regardless, this chapter was mostly about
the various new MapKit abilities available. It was also
interesting to see that protocols can also enforce you having
a certain property (in this case, a CLLocationCoordinate2D).