Complete Project 20
Completed Project 20. I will likely do the challenge to add the
score label soon, but as of right now I will hold off on the
challenges. The project was mostly about SpriteKit again and
introduced some new methods such as SKAction.follow, which
allows us to follow along a given CGPath (which we created
with UIBezierPath).