Commit Project 21
Completed Project 21 which revolved around local push
notifications. This is a really interesting area since I have seen
things like this discussed with the Apollo Reddit app and how
push notifications cost money (server processing fees) versus
the “less often but on device” local notifications. While it was
a really basic implementation, it was cool to get oriented with
UNUserNotificationCenter and all of the various methods.