Finish Project 22
Finish project 22. I haven’t been able to test this since I do not
have the updated iOS beta profile installed yet, but it was
interesting to learn about iBeacons and their potential. I am
not 100% sure when I would actually use one, but it is an
interesting concept nonetheless.