Finish Project 23
Finished Project 23, which was another SpriteKit game. It was pretty interesting to see
how we could implement rotation of images so easily with angularVelocity on the space
debris, but all in all I am not hugely interested in the SpriteKit games so I will be moving
onto the next project :).