Complete Project 25
Complete Project 25, which was all about the multipeer connectivity library. I never knew
about this library, but it opens up some interesting possiblities especially with some
things going on at work. I had no idea there was a library available that allowed for device
to device communication so it is nice to do a test app with it.