Complete Project 26
Complete Project 26, which was creating the classic “marble maze” game. While I don’t
usually like games in this series, it was cool being able to make a classic game that I
remember really impressing me when I first picked up iOS on my original iTouch. I’m not
sure if CoreMotion was around at the time (probably?) and SpriteKit probably was not,
so it is amazing to think how the original programmers must of had to hack around to get things to work. The CoreMotion library was probably the most interesting portion of this.
While I can’t see myself using it very much in normal work, it is nice to know it is there in
case I want to implement some really funky gestures or something :).