Complete Project 27
Completed Project 27, which was all about CoreGraphics. It was pretty interesting to see
all of the ways you can draw various shapes with CoreGraphcs. I am not sure when I will
be drawing directly to the screen, but I suppose something like a PDF annotation app
would be a great use for it by drawing shapes over the window. With that being said,
I am not sure how they would go about saving the images to the PDF (but maybe that
has something to do with the library).