Complete Project 28
Completed project 28 which was all about the LocalAuthentication library as well as
the third party keychain wrapper library. It was really cool to see how easy it was to use
TouchID with the evaluatePolicy method. It was also interesting seeing the keychain
code, but I think I would like to see how to actually use it beyond using a third party
library. While I think the wrapper will probably be around for a while, learning the actual
code seems to be the best way forward. Hopefully covered in Pro Swift!