Complete Project 30
Complete Project 30 which was all about Instruments. I enjoyed learning more about
Instruments, but I wasn’t a huge fan of the project itself. With that being said, it was
great to finally use Instruments because I have heard about them for a long time and
have never really had a project to use them on.