Commit UI of Project 31
Commited the Storyboard UI for Project 31. I didn’t have a ton of time to do any Hacking
With Swift today, but I wanted to make sure to do a little bit of work. I have used
StackViews a ton during my iOS development career, but learning about the usefulness
of the “pin” menu where I can pin to neighbors was a revelation. I plan on learning more
from “Modern AutoLayout”, but hopefully the rest of this chapter will be good for some
other tidbits I had not picked up in my use of StackViews so far.