Finish Project 31
Finished Project 31, which was all about utilizing a StackView and iPad multitasking features. While I had known a lot about StackViews before, the traitCollectionDidChange
method was a new one for me. I had been asked in an interview before how to make an
adaptive interface based on screen size and this would have helped a lot. Aside from
that, a lot of the stuff was a welcomed review.