Commit Project 32
Commit Project 32, which sets up initial UI for the project. Unfortunately I didn’t have a
lot of time for side projects today, but one thing I did learn from this project was how
easy it is to make autoresizing table view cells. I didn’t realize by putting the text label to
have 0 lines it would automatically resize to fit everything, so that is a very convenient
feature. With that being said, I am excited to try out using Dynamic Type because I have
heard a ton about it and how easy it is to make resizable text.