Finish Project 32
Finally finished Project 32, which was mostly about spotlight and how to index and
deindex with it. It was really cool to once again see how simple it was to add a really
cool feature to your app (and one that I use all the time with Spotlight). I also learned
about SFSafariViewController, which I wish a lot more apps would utilize since it uses
content blocking (unlike WKWebView).