Add functionality to pull data from CloudKit
Added functionality to pull data down from CloudKit using
CKQuery and CKQueryOperation. I also seem to have an issue with
a phantom TableViewCell that I will debug later, but I also changed
the code around to allow for the phantom cell to be created so the
app can load content from CloudKit and not crash from implicitly
unwrapping an optional.