Added fetching from CloudKit and more UI + bug fix
Added a view to write suggestions for each whistle and send them
up to CloudKit for storage. Also utilizes a CKRecord.Reference to
store a reference to the given whistle so if the whistle is deleted,
the suggestion is deleted from CloudKit as well. Finally, added in
functionality into the main whistle screen to redirect to the new UI
when a whistle is selected from the TableView (which also allows
for downloading and playing of the whistles).

I also fixed the bug of the “phantom whistle”. It was due to how I
was defining the newWhistles list as [Whistle()] which was
creating a nil Whistle contained in a list. I meant to put [Whistle]()
and when I did it fixed the issue.