Finish Project 33
Wow… That was a very long project. We did a lot with things such
as Push Notifications through CloudKit, more practice with
UserDefaults by saving what genres we wanted push notifications
for, and wrapping it all up with more UI work. All in all, a really long
but pretty satisfying project. I might end up doing the extras on
this one because a “pull to refresh” is very much needed for the
initial page.