Created Player class and fixed ReadMe
Finally back to doing some side project work! I ended up fixing a bug
with my nextEmptySlot method as well as added a class called Player
to represent the two players in the game. I also added a bit of logic to
the view portion of the app.

Also fixed an embarassing typo with my README file calling the book
“Hacking For Swift”. Whoops…