Finish two player mode and start AI
Finished up the two player mode so you can play with a friend. Also
began the implementation of the first part of the AI with GameplayKit.
So far I just created a Move class that helps GameplayKit score various
moves via a heuristic (conforming to GKGameModelUpdate). I also
started to conform the Board class to GKGameplayModel, but haven’t
finished the full implementation yet.