Finish Project 34
This project ended up being pretty cool because of the use of
GameplayKit. While the app itself is pretty simple, it was a great intro
into the idea of MVC because you essentially needed a model to pass
to the AI as well as a view to display to the user (and controller
functions to interact between both layers). As mentioned before, it also
introduced a lot of new GameplayKit protocols for implementing AI in
a game. I used GKGameModel to represent the game state (the Board),
I used GKGameModelUpdate to represent a move that could be made
by the AI (Move) and finally used GKGameModelPlayer to represent a
given player in the game (Player). Combining all of these with a
minimax strategy ended up giving me a very basic AI.

In the future, I would love to improve the heuristic on the AI so instead
of evaluating moves purely on if you win or lose, you could instead play
to win. Connect 4 is evidently a solved game, so it would be possible to
implement a really great AI. We will see if I can make that a weekend
project in the future!