Complete Project 36
Completed project 36, which was a flappy bird clone. This was another
project that was pretty much just all GameplayKit and SpriteKit, so it
wasn’t a whole lot of learning new iOS concepts. Making the games
can be fun, but I would just prefer to use things I could use at work.
Oh well!

This project introduced some background music via SKAudioNode and
also brought back the collision system via naming the nodes. Other
than that, we had some new implementation for gameplay (restaring,
logos, transitions, etc) but for the most part it was a lot of the same
stuff we have done before.