Start Project 37
Started project 37, which is an iPhone game utilizing the Apple Watch.
This seems like a pretty cool project since I have never made an Apple
Watch app before, so I am happy even though this is a game. We also
got a small taste of making custom delegates, which is an interesting
experience that I have used in work projects before. The rest of the
code is mostly an exercise in properly setting up the project and doing
some basic interface builder actions.