Finish Project 38
Finished Project 38, which was all about CoreData. Wow, it is a lot to
learn about. I know there are entire books on this, but I honestly feel
like I need to run through this with my own project to get the hang of
it completely. With that being said, it was a great primer of the native
database technology available for iOS. I learned about things like
NSFetchRequest, which queries your CoreData model for results. I also
learned about NSSortDescriptors, which allow you to sort your data
by given keys. There is also NSPredicate, which allows you to filter out
results from your fetch request. Finally, there is also
NSFetchedResultsController which is a container for your fetched
results and gives you multipe helper methods to help with managing
your data.