Complete Project 39
Completed the final project in Hacking With Swift. Woo! I learned a lot
about testing in this chapter, which was very welcomed since I really
don’t ever use tests during work. I would love to bring in some of these
tests to help out QA when running regression. Two other things I
leared about were the NSCountedSet (which can collect objects in an
array and count how many of each there are) and private setters (being
able to define something as private in the setting sense, but not in the
getting sense).